backend/requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
python-multipart==0.0.9
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
sqlalchemy==2.0.35
chromadb==0.5.15
sentence-transformers==3.1.1
pymupdf==1.24.10
python-docx==1.1.2
python-pptx==1.0.2
httpx==0.27.2
aiofiles==24.1.0
pydantic-settings==2.5.2
